Golang: Add error handling to all API functions.
The `morph` tool now post processes the SWIG generated code such that API
functions or methods by default now all return an error.  This is desirable
because it forces a tighter association between the code that generated the
error and the error itself.

This is achieved by injecting a call to cpl.ErrorReset()at the beginning of the
function, and cpl.LastError() is called when the function returns.

Instead of returning an error a function can instead be configured to panic when
an error is encountered by passing the `-panic-errors` option to morph.  This is
useful for functions that are not expected to normally generate errors and where
error handling can obfuscate the flow of code (e.g. getter methods).

Error handling in a function can be skipped entirely by passing the `-no-errors`
option to morph.